fix: resolve all remaining linting issues and warnings
- Fix fast refresh warnings in navigation-menu.tsx and sidebar.tsx by separating non-component exports
- Create navigation-menu-styles.ts for navigationMenuTriggerStyle CVA configuration
- Create sidebar-context.ts for useSidebar hook and context exports
- Create sidebar-styles.ts for sidebarMenuButtonVariants CVA configuration
- Fix useEffect missing dependency warning in CashuWallet.tsx
- All ESLint warnings now resolved (0 warnings, 0 errors)
- Project builds successfully with npm run ci
